+++ src/lib/st-table/shared/st-table-row/st-table-row.component.spec.ts
@@ -12,6 +12,8 @@ import { CommonModule } from '@angular/common';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { StTableRowComponent } from './st-table-row.component';
+import { By } from '@angular/platform-browser';
+import { ChangeDetectionStrategy } from '@angular/core';
 
 let fixture: ComponentFixture<StTableRowComponent>;
 let component: StTableRowComponent;
@@ -21,6 +23,10 @@ describe('StTableRowComponent', () => {
       TestBed.configureTestingModule({
          imports: [CommonModule, RouterTestingModule],
          declarations: [StTableRowComponent]
+      })
+         // remove this block when the issue #12313 of Angular is fixed
+         .overrideComponent(StTableRowComponent, {
+         set: {changeDetection: ChangeDetectionStrategy.Default}
       })
          .compileComponents();  // compile template and css
    }));
@@ -87,5 +93,22 @@ describe('StTableRowComponent', () => {
 
          expect(fixture.nativeElement.classList).toContain('selected');
       });
+   });
+
+   it('hover menu has to be hidden if developer does not add any content to it', () => {
+      let hoverMenu: Element = fixture.nativeElement.querySelector('.hover-menu');
+      let contentElement: HTMLSpanElement = document.createElement('span');
+
+      hoverMenu.appendChild(contentElement);
+      component.ngOnInit();
+      fixture.detectChanges();
+
+      expect(fixture.debugElement.query(By.css('.hover-menu')).properties.hidden).toBe(false);
+
+      hoverMenu.removeChild(contentElement);
+      component.ngOnInit();
+      fixture.detectChanges();
+
+      expect(fixture.debugElement.query(By.css('.hover-menu')).properties.hidden).toBe(true);
    });
 });